+++ assets/javascripts/kitten/karl/pages/kissbankers-modal/stories.js
@@ -1,0 +1,7 @@
+import React from 'react'
+import { storiesOf } from '@storybook/react'
+import KissbankersModal from './index'
+
+storiesOf('Pages/KissbankersPage', module).add('Kissbankers-modal', () => {
+  return <KissbankersModal />
+})